Separated test from actual code
