+++ scaffold/test/factories/{{{scaffold_factory}}}.js
@@ -1,6 +1,6 @@
 import { factory } from 'factory-girl';
 
-import {{{scaffold_entity_capitalise}}} from '../../src/persistence/mongoose/User/model'
+import {{{scaffold_entity_capitalise}}} from '../../src/persistence/mongoose/{{{scaffold_entity_capitalise}}}/model'
 
 export default factory.define('{{{scaffold_entity_capitalise}}}', {{{scaffold_entity_capitalise}}}, {
   name: 'test'
